Moved authentication to application-wide configuration and moved password dehash/check to user-controller.
