+++ controllers/user-controller.js
@@ -3,44 +3,6 @@ const passport = require('passport'),
       bcrypt = require('bcryptjs'),
       user_model = require('../models/user_model');
 
-passport.use(new LocalStrategy(
-    function(username, password, done){
-        let user_login = new user_model();
-        user_login = user_controller.load_by_username(username);
-
-        // Check if user was found, will be null for username if not.
-        if (!user_login.username.value){
-            return done(null, false, {message: 'Unable to find account associated.'});
-        }else{
-            // User found, compare passwords
-            // Put in user model?
-            console.log("INFO RECEIVED:\n\tUsername:\t"+username+"\n\tPassword:\t"+password);
-            let password_match = bcrypt.compare(password, user_login.password.value, function(err, isMatch){
-                if (err){
-                    console.log("Error encountered: " + err);
-                }else{
-                    if (isMatch){
-                        return done(null, user_login);
-                    }else{
-                        return done(null, false, {message: 'Password does not match.'});
-                    }
-                }
-            });
-        }
-    }
-));
-
-passport.serializeUser((user, done) => {
-    done(null, user.id.value);
-});
-
-passport.deserializeUser((id, done) => {
-    const new_user_model = new user_model();
-    new_user_model.schema.load(id);
-    let err = (new_user_model.schema.id.value == null) ? "User not found!" : null;
-    done(err, new_user_model);
-});
-
 let user_controller = {
     login: (req,res) => {
         console.log(req.body);
@@ -119,6 +81,11 @@ let user_controller = {
     create_user: function(username, hash, files = {}) {
       let create_user_instance = new user_model();
       create_user_instance.set(null, username, hash, {});
+    },
+    password_compare: function(password, comparePassword, callback){
+        let password_match = bcrypt.compare(password, comparePassword, function(err, isMatch){
+            callback(err, isMatch);
+        });
     }
 }
